Improved Selection Behavior.
The `selected_rows`  variable now holds row IDs instead of indices, So the same things stay selected even if the rows are moved around.

There is now a multiple selection ability. Hold the CTRL key to make a range selection; Hold the SHIFT key to select and unselect specific rows.

Rows will be unselected if they are hidden.